requirements.txt
langchain==0.0.189
streamlit==1.23.1
python-dotenv==1.0.0
pypdf2==3.0.1
openai==0.27.7
tiktoken==0.4.0
faiss_cpu==1.7.4